Add AMF and QoS billing to web portal invoice views
The web portal invoice page (/nodes/billing/invoice/YYYY/MM/) and its
JSON export previously showed only reservations.  The REST API endpoint
(/nodes/api/invoice/YYYY/MM/) included AMF and QoS data since PR #63,
but the portal views were not updated at that time because they are
separate classes in a different file (views/billing.py vs api/views.py).

This commit:

1. Extracts shared billing logic into utils/invoice_builders.py:
   - build_reservation_lines(), build_amf_lines(), build_qos_lines()
   - build_combined_response(), get_project_cost_breakdown()
   - Reservation hour/cost helpers

2. Refactors InvoiceReportView (API) to delegate to shared helpers
   instead of using private methods.  Sub-endpoints
   (InvoiceReservationsView, InvoiceAMFView, InvoiceQoSView) now
   inherit from APIView directly and call the shared functions.

3. Updates InvoiceDetailView.get_context_data() to call
   build_amf_lines() and build_qos_lines(), merging entries into the
   per-project grouping alongside reservation lines.

4. Updates InvoiceExportView to include amf_entries and qos_entries
   per project, and total_amf_entries/total_qos_entries in metadata.

5. Updates invoice_detail.html template:
   - Header shows AMF/QoS counts alongside reservation count
   - Each project card includes AMF and QoS sub-tables when present
   - Empty state changed from "No completed reservations" to
     "No billable activity"

6. Updates developer docs (invoice-views.md, CHANGELOG.md).
